requirements.txt
fastapi
uvicorn[standard]
python-multipart
starlette
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
annotated-types
anyio
appdirs
asttokens
attrs
beautifulsoup4
botocore
cachetools
certifi
charset-normalizer
colorama
comm
contourpy
cycler
dataclasses-json
debugpy
decorator
distro
duckdb
executing
fastapi
filetype
fonttools
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
html5lib
httpcore
httplib2
httpx
httpx-sse
idna
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jedi
jiter
jmespath
joblib
jsonpatch
jsonpointer
jupyter_client
jupyter_core
kiwisolver
langchain
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
marshmallow
matplotlib
matplotlib-inline
multidict
mypy_extensions
nest-asyncio
numpy
openai
orjson
ormsgpack
packaging
pandas
parso
pillow
platformdirs
prompt_toolkit
propcache
proto-plus
protobuf
psutil
pure_eval
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
pyee
Pygments
pyparsing
pyppeteer
python-dateutil
python-dotenv
pytz
pywin32
PyYAML
pyzmq
regex
requests
requests-toolbelt
rsa
s3fs
scikit-learn
scipy
seaborn
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
tenacity
threadpoolctl
tiktoken
tornado
tqdm
traitlets
typing-inspect
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
wcwidth
webencodings
websockets
wrapt
xxhash
yarl
zipp
zstandard
duckdb
boto3
seaborn